requirements.txt
asgiref==3.7.2
Django==5.0.3
psycopg2-binary==2.9.1
sqlparse==0.4.4
pillow==10.3.0
python-dotenv==1.0.1
requests==2.31.0